+++ djrill/tests/test_mandrill_subaccounts.py
@@ -7,40 +7,36 @@ from .mock_backend import DjrillBackendMockAPITestCase
 class DjrillMandrillSubaccountTests(DjrillBackendMockAPITestCase):
     """Test Djrill backend support for Mandrill subaccounts"""
 
-    def test_send_basic(self):
-        mail.send_mail('Subject here', 'Here is the message.',
-            'from@example.com', ['to@example.com'], fail_silently=False)
-        self.assert_mandrill_called("/messages/send.json")
+    def test_no_subaccount_by_default(self):
+        mail.send_mail('Subject', 'Body', 'from@example.com', ['to@example.com'])
         data = self.get_api_call_data()
-        self.assertEqual(data['message']['subject'], "Subject here")
-        self.assertEqual(data['message']['text'], "Here is the message.")
-        self.assertFalse('from_name' in data['message'])
-        self.assertEqual(data['message']['from_email'], "from@example.com")
-        self.assertEqual(len(data['message']['to']), 1)
-        self.assertEqual(data['message']['to'][0]['email'], "to@example.com")
         self.assertFalse('subaccount' in data['message'])
 
-    @override_settings(MANDRILL_SUBACCOUNT="test_subaccount")
-    def test_send_from_subaccount(self):
-        mail.send_mail('Subject here', 'Here is the message.',
-            'from@example.com', ['to@example.com'], fail_silently=False)
-        self.assert_mandrill_called("/messages/send.json")
+    @override_settings(MANDRILL_SETTINGS={'subaccount': 'test_subaccount'})
+    def test_subaccount_setting(self):
+        mail.send_mail('Subject', 'Body', 'from@example.com', ['to@example.com'])
         data = self.get_api_call_data()
-        self.assertEqual(data['message']['subject'], "Subject here")
-        self.assertEqual(data['message']['text'], "Here is the message.")
-        self.assertFalse('from_name' in data['message'])
-        self.assertEqual(data['message']['from_email'], "from@example.com")
-        self.assertEqual(len(data['message']['to']), 1)
-        self.assertEqual(data['message']['to'][0]['email'], "to@example.com")
         self.assertEqual(data['message']['subaccount'], "test_subaccount")
 
-    @override_settings(MANDRILL_SUBACCOUNT="global_setting_subaccount")
+    @override_settings(MANDRILL_SETTINGS={'subaccount': 'global_setting_subaccount'})
     def test_subaccount_message_overrides_setting(self):
-        message = mail.EmailMessage(
-            'Subject here', 'Here is the message',
-            'from@example.com', ['to@example.com'])
+        message = mail.EmailMessage('Subject', 'Body', 'from@example.com', ['to@example.com'])
         message.subaccount = "individual_message_subaccount"  # should override global setting
         message.send()
-        self.assert_mandrill_called("/messages/send.json")
+        data = self.get_api_call_data()
+        self.assertEqual(data['message']['subaccount'], "individual_message_subaccount")
+
+    # Djrill 1.x offered dedicated MANDRILL_SUBACCOUNT setting.
+    # In Djrill 2.x, you should use the MANDRILL_SETTINGS dict as in the earlier tests.
+    # But we still support the old setting for compatibility:
+    @override_settings(MANDRILL_SUBACCOUNT="legacy_setting_subaccount")
+    def test_subaccount_legacy_setting(self):
+        mail.send_mail('Subject', 'Body', 'from@example.com', ['to@example.com'])
+        data = self.get_api_call_data()
+        self.assertEqual(data['message']['subaccount'], "legacy_setting_subaccount")
+
+        message = mail.EmailMessage('Subject', 'Body', 'from@example.com', ['to@example.com'])
+        message.subaccount = "individual_message_subaccount"  # should override legacy setting
+        message.send()
         data = self.get_api_call_data()
         self.assertEqual(data['message']['subaccount'], "individual_message_subaccount")