+++ docs/history.rst
@@ -96,6 +96,9 @@ Other Djrill 2.0 Changes
   (You can also directly manage your own long-lived Djrill connection across multiple sends,
   by calling open and close on :ref:`Django's email backend <django:topic-email-backends>`.)
 
+* Add global :setting:`MANDRILL_SETTINGS` dict that can provide defaults
+  for most Djrill message options.
+
 * Add :exc:`djrill.NotSerializableForMandrillError`